backend/requirements.txt
torch
torchvision
timm
facenet-pytorch
fastapi
uvicorn[standard]
python-multipart
opencv-python
Pillow
scikit-learn
matplotlib
seaborn
aiosqlite
numpy
requests
